ox-odt.el: Fix handling of ODT attributes
* lisp/ox-odt.el (org-odt--read-attribute): New.
(org-odt-template, org-odt-link--inline-image, org-odt-src-block)
(org-odt-table-style-spec, org-odt-table-cell)
(org-odt--translate-list-tables): Use it.

----------------------------------------------------------------

To the extent possible under law, the author(s) have dedicated
all copyright and related and neighboring rights to this software
to the public domain worldwide. This software is distributed
without any warranty.

You should have received a copy of the CC0 Public Domain
Dedication along with this software. If not, see

   <http://creativecommons.org/publicdomain/zero/1.0/>.

----------------------------------------------------------------