lisp/ox-odt.el: Improve support for pagebreaks
* etc/styles/OrgOdtStyles.xml (OrgPageBreakDefault)
(OrgPageBreakLandscape):  Remove it.

* lisp/ox-odt.el (org-odt-page-break-style-format): New constant.
(org-odt--get-derived-paragraph-style): New helper routine.
(org-odt--read-attribute): Improve.
(org-odt-keyword): Invert how PAGEBREAKs are handled.  Now,
PAGEBREAK keywords create an empty paragraph on the current
page, which is then *followed* with a new page.  (Previously,
this keyword created a new page, in which an empty paragraph
was created.)
(org-odt-paragraph, org-odt-headline, org-odt-special-block):
Use derived paragraph styles that honor the three new
properties: `:page-break', `:page-style', and `page:number'.
`:page-break' can take one of the following values: "before",
t, "after".  These properties correspond with the options
under Paragraph->Text Flow->Breaks. See
https://help.libreoffice.org/Writer/Text_Flow#Breaks.)

a) To shift a paragraph to a new page, use
      #+ATTR_ODT: :pagebreak t
      Blah blah blah blah ...

b) To shift a paragraph to a new page, that uses a specific
page style, say "OrgFirstPage", use

      #+ATTR_ODT: :page-style "OrgFirstPage"
      Blah blah blah blah ...

c) To shift a paragraph to a new page, that uses a specific
page style, say "OrgFirstPage", and to begin numbering the
page from numeral 1, use

      #+ATTR_ODT: :page-style "OrgFirstPage" :page-number 1
      Blah blah blah blah ...

The above properties can also be applied to headlines.  For
example, to shift a headline to a new page, that uses a
specific page style, say "OrgFirstPage", and to begin
numbering the page from numeral 1, use

      * Copyright Statement
      :PROPERTIES:
      :ATTR_ODT: :page-number 1 :page-style "OrgFirstPage"
      :END:

See #29.